+++ lib/ui/schedule/view_model/calendars/week_viewmodel.dart
@@ -2,7 +2,7 @@ import 'package:calendar_view/calendar_view.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 
 import 'package:notredame/utils/utils.dart';
-import 'calendar_viewmodel.dart';
+import '../calendar_viewmodel.dart';
 
 class WeekViewModel extends CalendarViewModel {
   // Sunday of current week
